finished updating documentation and progress on improving pathfinding - only one or two cases where shortest path is not found

+++ app/src/bham/bioshock/common/pathfinding/PathfindingValues.java
@@ -9,10 +9,10 @@ public class PathfindingValues {
     private int pathCost;
 
     // the heuristic estimate of the cost to get to the goal point
-    private int heuristicCost;
+    private double heuristicCost;
 
     // the total cost (pathCost + heuristicCost) to go from this node to the goal node
-    private int totalCost;
+    private double totalCost;
 
     // the parent point to this point - the point that came before
     private Coordinates parent;
@@ -45,12 +45,12 @@ public class PathfindingValues {
     }
 
     // method to get the heuristicCost
-    public int getHeuristicCost() {
+    public double getHeuristicCost() {
         return heuristicCost;
     }
 
     // method to get the totalCost
-    public int getTotalCost() {
+    public double getTotalCost() {
         return totalCost;
     }
 
@@ -70,12 +70,12 @@ public class PathfindingValues {
     }
 
     // method to set the heuristicCost
-    public void setHeuristicCost(int heuristicCost){
+    public void setHeuristicCost(double heuristicCost){
         this.heuristicCost = heuristicCost;
     }
 
     // method to set the totalCost
-    public void setTotalCost(int totalCost){
+    public void setTotalCost(double totalCost){
         this.totalCost = totalCost;
     }
